yaml encoding didn't work for long sized integers.
changed the Json.Num case in jsonToYaml to use BigDecimal.longValue instead of intValue

added a test

+++ zio-json-yaml/src/main/scala/zio/json/yaml/package.scala
@@ -126,7 +126,7 @@ package object yaml {
       case Json.Num(value) =>
         val stripped = value.stripTrailingZeros()
         if (stripped.scale() <= 0) {
-          new ScalarNode(Tag.INT, stripped.intValue().toString, null, null, options.scalarStyle(json))
+          new ScalarNode(Tag.INT, stripped.longValue.toString, null, null, options.scalarStyle(json))
         } else {
           new ScalarNode(Tag.FLOAT, stripped.toString, null, null, options.scalarStyle(json))
         }
